requirements.txt
chainlit==2.5.5
openai==1.76.0
python-dotenv==1.1.0
PyMuPDF==1.25.5
langchain

